Refactored ExerciseDetail view with new ExerciseChart component that encapsulates all logic rela ted to the 1RM/volume chart
