+++ client/src/features/exercise/ExerciseDetail.tsx
@@ -1,20 +1,19 @@
 import { useParams } from "react-router-dom";
 import { useQuery } from "@tanstack/react-query";
-import { useEffect, useState, type FormEvent } from "react";
-import { Line } from "react-chartjs-2";
+import { useEffect, useState } from "react";
 import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend} from 'chart.js';
 import { useAuth } from "../../context/AuthContext";
 import { getCategoryLabel } from "../../shared/utils/constants/category";
-import type { Exercise, ExerciseTimeseries } from "../exercises/exercises.types";
+import type { Exercise } from "../exercises/exercises.types";
 import type { ExerciseLog } from "../workout/workout.types";
 import type { Params } from "react-router-dom";
 import type { columnConfig } from "../../shared/components/Table/table.types";
-import type { ChartData, Point } from "chart.js";
 import ContentSection from "../../shared/components/ContentSection";
 import EventMessage from "../../shared/components/EventMessage";
 import KpiCard from "../../shared/components/KpiCard";
 import Table from "../../shared/components/Table";
 import ExerciseEditPopup from "./views/ExerciseEditPopup";
+import ExerciseChartComponent from "./views/ExerciseChart";
 import './exerciseDetail.css'
 
 
@@ -24,7 +23,6 @@ function ExerciseDetail(){
     const {axios_instance} = useAuth()!
     const [logs,setLogs] = useState<Array<ExerciseLog>|null>(null)
     const [metrics, setMetrics] = useState<exerciseMetrics|null>(null)
-    const [chartData, setChartData] = useState<ChartData<"line", (number | Point | null)[], unknown>|null>(null)
     const [editingExercise, setEditingExercise] = useState<boolean>(false)
 
     ChartJS.register( CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend);
@@ -47,45 +45,6 @@ function ExerciseDetail(){
         {key:'distance', header:'Distance (km)'},
         {key:'duration', header:'Duration (mins)'},
     ]
-
-
-    // Chart configuration
-    const options = {
-        responsive: true,
-        interaction: {
-            mode: 'index' as const,
-            intersect: false,
-        },
-        stacked: false,
-        // plugins: {
-        //     title: {
-        //     display: true,
-        //     text: 'Chart.js Line Chart - Multi Axis',
-        //     },
-        // },
-        scales: {
-            y: {
-                type: 'linear' as const,
-                display: true,
-                position: 'left' as const,
-            },
-            y1: {
-                type: 'linear' as const,
-                display: true,
-                position: 'right' as const,
-                grid: {
-                    drawOnChartArea: false,
-                },
-            },
-            x:{
-                display:false,
-                ticks: {
-                    minRotation: 0,
-                    maxRotation: 0
-                }
-            }
-        },
-    };
 
     // Data fetch, with query_params 
 
@@ -117,59 +76,7 @@ function ExerciseDetail(){
                 'associated workouts': [exercise.kpis.associated_workouts ?? 0, 0],
                 'total logs': [exercise.kpis.total_logs ?? 0, 0],
                 'logs current month': [exercise.kpis.logs_current_month ?? 0,0]
-            })
-            const timeseries:ExerciseTimeseries|null = exercise.timeseries
-            if(timeseries){
-                const labels:string[] = (Object.keys(timeseries))
-                if(exercise.category == 'str'){
-                    setChartData({
-                    labels,
-                    datasets: [
-                        {
-                            label: '1RM',
-                            data: labels.map((label:string)=> {
-                                return timeseries[label]['1RM']
-                            }),
-                            borderColor: 'rgb(250 75 42)',
-                            
-                            yAxisID: 'y'
-                        },
-                        {
-                            label: 'Volume',
-                            data: labels.map((label:string)=> {
-                                return timeseries[label]['volume']
-                            }),
-                            borderColor: 'rgb(215 243 31)',
-                            yAxisID: 'y1'
-                        },
-                    ],
-                    })
-                } else if (exercise.category == 'car'){
-                    setChartData({
-                    labels,
-                    datasets: [
-                        {
-                            label: 'Distance',
-                            data: labels.map((label:string)=> {
-                                return timeseries[label]['distance']
-                            }),
-                            borderColor: 'rgb(250 75 42)',
-                            
-                            yAxisID: 'y'
-                        },
-                        {
-                            label: 'Duration (mins)',
-                            data: labels.map((label:string)=> {
-                                return timeseries[label]['duration']
-                            }),
-                            borderColor: 'rgb(215 243 31)',
-                            yAxisID: 'y1'
-                        },
-                    ],
-                    })
-                }
-                
-            }
+            })          
         }
     },[exercise])
 
@@ -223,9 +130,10 @@ function ExerciseDetail(){
         <div className="row mb-4">
             <div className="col-12">
                 <ContentSection title="Timeseries metrics">
-                    {chartData ?
+                    {/* {chartData ?
                     <Line data={chartData} options={options}></Line>
-                    :<EventMessage style="warning" message="At least two exercise logs are required to track these metrics."/>}
+                    :<EventMessage style="warning" message="At least two exercise logs are required to track these metrics."/>} */}
+                    <ExerciseChartComponent timeseries={exercise.timeseries} category={exercise.category} />
                 </ContentSection>
             </div>
         </div>
